Start a graphical user interface (GUI) based on the gWidgetsWWW2 toolkit.

This function starts a browser based GUI. Please visit the github page of gWidgetsWWW2 for an explanation how this toolkit works.

gmkin(script_name, show.log = FALSE)

Arguments

script_name
During development, a script name with a local working version of gmkin can be passed. Defaults to the location of the gmkin.R script shipped with the package.

show.log
During development, it may be useful to see the log of the Rook apps.

Value

The function is called for its side effect, namely starting the GUI in a browser. For the curious, the desperate or the adventurous, the gmkin app (a GWidgetsApp object) is returned.

Examples

  ## Not run: gmkin() # Start the gmkin GUI
  not_run({
      g <- gmkin()

      # Only for debugging or other advanced usage:
      session <- gmkin:::get_current_session(g)
      session$p.df
      session$ds.df
      session$ds.cur
      session$m.cur
      session$ws

})